+++ bootstrap/bootstrap.sh
@@ -3,7 +3,7 @@
 cat /etc/resolv.conf
 
 apt-get update -y 
-apt-get install -y containernetworking-plugins containerd tcpdump curl inetutils-ping iptables fuse-overlayfs procps bash iproute2
+apt-get install -y dnsmasq containernetworking-plugins containerd tcpdump curl inetutils-ping iptables fuse-overlayfs procps bash iproute2 dnsutils net-tools
 mkdir -p /etc/cni/net.d
 
 export KUBERNETES_VERSION=v1.21.1
@@ -15,6 +15,7 @@ export KUBECONFIG=$MESOS_SANDBOX/kubeconfig.yaml
 update-alternatives --set iptables /usr/sbin/iptables-legacy
 curl -sfL https://get.k3s.io | sh -
 curl https://raw.githubusercontent.com/AVENTER-UG/go-mesos-framework-k3s/master/bootstrap/dashboard_auth.yaml > $MESOS_SANDBOX/dashboard_auth.yaml
+curl https://raw.githubusercontent.com/AVENTER-UG/go-mesos-framework-k3s/master/bootstrap/dashboard_traefik.yaml > $MESOS_SANDBOX/dashboard_traefik.yaml
 curl https://raw.githubusercontent.com/kubernetes/dashboard/v2.2.0/aio/deploy/recommended.yaml > $MESOS_SANDBOX/dashboard.yaml
 if [ "$K3SFRAMEWORK_TYPE" = "server" ]
 then